Edit script.js
Modified the display functions to display the current number the
user is inputting.

The calculator strings together several operations, with each pair
of numbers being evaluated at a time. Also, if the user inputs double
operators (either similar or different), the operand is doubled or
divided by itself depending on the first operator.