Fix bugs
Fixed the issue where an error was being thrown if any operator
button was pressed before inputting a number.

Answers with long decimals are now rounded to prevent screen overflow.

Added a function that removes the previous calculated solution if
the user starts a number input before pressing an operator.